still working on bullet collisions...
there's quite a bit of work here, and nothing super fruitful yet. i'm
thinking the real way that we should do this, because if you just do "is
my current point outside of my polygon", you could totally miss it and
skip right over your shape, so for the moment we have to work on a
predictive model. there's a large comment in 'WillCollide' about what
the deal is with this, so read that if you got to the end of this and
said "????"